Toggle fullscreen mode with F11, abstract window init and functions
