+++ CMakeLists.txt
@@ -11,7 +11,7 @@ add_subdirectory(opl)
 set(TITUS_SOURCES
     opentitus.cpp
     src/audio.cpp
-    src/backbuffer.cpp
+    src/window.cpp
     src/common.cpp
     src/draw.cpp
     src/enemies.cpp
