+++ opentitus.cpp
@@ -38,27 +38,14 @@
 #include "settings.h"
 #include "malloc.h"
 #include "sprites.h"
-#include "backbuffer.h"
+#include "window.h"
 #include "viewimage.h"
 #include "fonts.h"
 #include "menu.h"
 #include "engine.h"
 #include "original.h"
 #include "objects.h"
-
-SDL_Window *window;
-SDL_Renderer *renderer;
-
-const char* getGameTitle() {
-    switch(game) {
-        case GameType::Titus:
-            return "OpenTitus";
-        case GameType::Moktar:
-            return "OpenMoktar";
-        default:
-            return "Something else...";
-    }
-}
+#include "window.h"
 
 int init() {
 
@@ -74,42 +61,10 @@ int init() {
     }
 
 
-    Uint32 windowflags = 0;
-    int w;
-    int h;
-    switch (videomode) {
-        default:
-        case 0: //window mode
-            w = 960;
-            h = 600;
-            windowflags = SDL_WINDOW_RESIZABLE;
-            break;
-        case 1: // fullscreen
-            w = 0;
-            h = 0;
-            windowflags = SDL_WINDOW_FULLSCREEN_DESKTOP;
-            break;
+    retval = initwindow();
+    if(retval != 0) {
+        return retval;
     }
-
-    window = SDL_CreateWindow(
-        getGameTitle(),
-        SDL_WINDOWPOS_UNDEFINED,
-        SDL_WINDOWPOS_UNDEFINED,
-        w,
-        h,
-        windowflags
-    );
-    renderer = SDL_CreateRenderer(window, -1, SDL_RENDERER_ACCELERATED);
-    if (renderer == NULL) {
-        printf("Unable to set video mode: %s\n", SDL_GetError());
-        return TITUS_ERROR_SDL_ERROR;
-    }
-
-    // screen = SDL_GetWindowSurface(window);
-    screen = SDL_CreateRGBSurfaceWithFormat(0, 320, 200, 32, SDL_GetWindowPixelFormat(window));
-
-    SDL_RenderSetLogicalSize(renderer, 320, 200);
-    SDL_SetRenderDrawColor(renderer, 0, 0, 0, 255);
 
     initaudio();
