+++ src/gates.cpp
@@ -30,7 +30,7 @@
 #include <stdlib.h>
 #include "SDL2/SDL.h"
 #include "globals.h"
-#include "backbuffer.h"
+#include "window.h"
 #include "sprites.h"
 #include "settings.h"
 #include "audio.h"
