+++ src/window.h
@@ -22,19 +22,20 @@
  * General Public License for more details.
  */
 
-/* backbuffer.h
- * Contains backbuffer pointers
+/* window.h
+ * Window and window back buffer pointers
 */
 
-#ifndef BACKBUFFER_H
-#define BACKBUFFER_H
-
+#pragma once
 #include "SDL2/SDL.h"
 
+int initwindow();
+void togglefullscreen();
+void togglefullscreendone();
+
 void SDL_Flip(SDL_Surface * screen);
+
 extern SDL_Window *window;
 extern SDL_Renderer *renderer;
 extern SDL_Surface *screen; //Backbuffer
 extern SDL_Surface *tilescreen; //Tile screen
-
-#endif
